+++ amd/__init__.py
@@ -1,8 +1,7 @@
 from .io.CifReader import CifReader
 from .io.CSDReader import CSDReader
 from .io.PeriodicSetIO import SetReader, SetWriter
-from .io._Reader import cellpar_to_cell
+from .core.PeriodicSet import PeriodicSet
 from .calculate import *
 from .compare import *
-from .core.PeriodicSet import *
-from .core.eta import *
+from .utils import *