+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.6] - 03/11/2021
+
+### Added
+
+- Optional parameter low_memory added to AMD comparison functions. Uses a slower but more memory efficient method, useful for larger collections of AMDs.
+
+- The SetReader now has .to_dict() which returns a dict containing the scalar (strings and numbers) data attached to PeriodicSets. These could either be attrs of the h5py group containing the PeriodicSet, or values stored in the .tags dict of the PeriodicSet. The format of the returned dict is easily passable to pandas.DataFrame().
+
+- To compliment the addition of .to_dict(), the CifReader and CSDReader now have ways to extract additional data from ccdc Entries (later ase CifBlocks). The parameter extract_data can be passed as a dictionary with data (column) titles as keys and callable functions as values, which take the Entry object and return the data. E.g. if the custom function get_density() takes the entry and returns the density, use extract_data={'Density': get_density}, then the PeriodicSets will have the tag 'Density' with a value given by get_density(entry). This approach is flexible to any function, but some common ones are included in amd.io.ccdc_utils.
+
+- CifReader and CSDReader accept include_if parameter, which can be None or a list of callables. This can be used to discard structures with specific properties. If include_if is a list, the Entry/CifBlock is passed to each callable and is skipped if any do not return True.
+
+### Changed
+
+### Removed
+
+- amd.compare() has been removed, as have the old lower case versions of functions amd.amd(), amd.amd_pdist(), etc.
+
+- types=True removed as optional parameter to Readers, types are now always read.
+
+- The option to pass a range of k values to comparison funcitons, along with the optional parameter ord, have been removed. Comparing over a range of k can be useful, but often its not if the range of distances aren't stored. No computational savings were really made in the function as opposed to a naive approach.
+
 ## [1.1.5] - 02/11/2021
 
 ### Added
@@ -13,8 +35,6 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - EMD has updated internals which is 2-3x faster.
-
-### Removed
 
 ## [1.1.4] - 13/09/2021
 
@@ -32,7 +52,7 @@ All notable changes to this project will be documented in this file.
 
 - The optional keyword argument types is no longer used for the Readers; types are read in regardless.
 
-## [1.1.1] - 18/06/2021
+## [1.1.1] - 18/07/2021
 
 ### Added